refactored io to discourage use of generics
