+++ ramannoodle/dynamics/phonon.py
@@ -5,7 +5,7 @@ from numpy.typing import NDArray
 
 from ramannoodle.dynamics.abstract import Dynamics
 from ramannoodle.globals import RAMAN_TENSOR_CENTRAL_DIFFERENCE
-from ramannoodle.polarizability import PolarizabilityModel
+from ramannoodle.polarizability.abstract import PolarizabilityModel
 from ramannoodle.spectrum.raman import PhononRamanSpectrum
